Integrated compress side

+++ szx/src/cuszx_float.cu
@@ -342,7 +342,7 @@ __global__ void compress_float(float *oriData, unsigned char *meta, short *offse
             int idx = 0;
             if (tidx ==0 && tidy == 0)
             {
-                idx = atomicAdd(&num_state2, num_sig);
+                idx = atomicAdd(&num_state2, (uint64_t)num_sig);
                 blk_idx[b] = idx;    // Store the index of where this block has values and indices within block
             }
             __syncthreads();
